New & updated plugins
- 6 new plugins (additional 4 in Additional Plugins folder)
 - All new plugins are portable programs that were also added to ATOM Store
- Simplified some logic in Programs-Hashtable.ps1
- Fixed issue where 7-Zip wouldn't launch out of temp folder
- Added auto-scrolling to end in OutputBox within ATOM Store, Neutron, and Detectron
 - Fixed typo that caused auto-scrolling to end to not work in ATOMizer
- Neutron changes
 - Winget and Chocolatey auto-install after clicking "Run" button if not detected (removed 'Install Winget' and 'Install Chocolatey' Customizations"
 - Added 20 new programs
 - Renamed "Winget Upgrade" to "Upgrade Apps"

+++ ATOM/Dependencies/Neutron/Functions/Install-Programs.ps1
@@ -5,6 +5,44 @@ function Install-Programs {
 		[System.Collections.ArrayList]$selectedInstallPrograms
 	)
 	
+	# Install Winget if not detected
+	$wingetPath = Join-Path $env:LOCALAPPDATA "Microsoft\WindowsApps\winget.exe"
+	$wingetExists = Test-Path $wingetPath
+	if (!$wingetExists) {
+		Write-OutputBox "Winget not detected. Installing..."
+		Set-PSRepository -Name 'PSGallery' -InstallationPolicy Trusted
+		Install-Script -Name winget-install -Force
+		$wingetArgument = "-ExecutionPolicy Bypass winget-install.ps1"
+		Start-Process powershell -ArgumentList $wingetArgument -Wait
+		
+		$wingetExists = Test-Path $wingetPath
+		if ($wingetExists) {
+			Write-OutputBox "- Installed Winget"
+		} else {
+			Write-OutputBox "- Failed to install Winget"
+		}
+		
+		Write-OutputBox ""
+	}
+	
+	# Install Chocolatey if not detected
+	$chocoPath = Join-Path $env:PROGRAMDATA "chocolatey\choco.exe"
+	$chocoExists = Test-Path $chocoPath
+	if (!$chocoExists) {
+		Write-OutputBox "Chocolatey not detected. Installing..."
+		$chocoArgument = "Set-ExecutionPolicy Bypass -Scope Process -Force; [System.Net.ServicePointManager]::SecurityProtocol = [System.Net.ServicePointManager]::SecurityProtocol -bor 3072; iex ((New-Object System.Net.WebClient).DownloadString('https://community.chocolatey.org/install.ps1'))"
+		Start-Process powershell -ArgumentList $chocoArgument -Wait
+		
+		$chocoExists = Test-Path $chocoPath
+		if ($chocoExists) {
+			Write-OutputBox "- Installed Chocolatey"
+		} else {
+			Write-OutputBox "- Failed to install Chocolatey"
+		}
+		
+		Write-OutputBox ""
+	}
+	
 	foreach ($selectedProgram in $selectedInstallPrograms) {
 
 		foreach ($category in $installPrograms.Keys) {
@@ -13,8 +51,8 @@ function Install-Programs {
 				$programInfo = $installPrograms[$category][$selectedProgram]
 				Write-OutputBox $selectedProgram
 				
-				# Try to install with winget
-				if ($programInfo['winget']) {
+				# Try to install with Winget
+				if ($programInfo['winget'] -and $wingetExists) {
 					$process = Start-Process -FilePath 'winget' -ArgumentList "install --id $($programInfo['winget']) --accept-package-agreements --accept-source-agreements --force" -Wait -PassThru
 					if ($process.ExitCode -ne 0) {
 						Write-OutputBox "- Failed to install with Winget"
@@ -24,8 +62,8 @@ function Install-Programs {
 					}
 				}
 
-				# If winget fails, try to install with Chocolatey
-				if ($programInfo['choco']) {
+				# If Winget fails, try to install with Chocolatey
+				if ($programInfo['choco'] -and $chocoExists) {
 					$process = Start-Process -FilePath 'choco' -ArgumentList "install $($programInfo['choco']) -y" -Wait -PassThru
 					if ($process.ExitCode -ne 0) {
 						Write-OutputBox "- Failed to install with Chocolatey"
@@ -35,7 +73,7 @@ function Install-Programs {
 					}
 				}
 
-				# If chocolatey fails, try to download and install from URL
+				# If Chocolatey fails, try to download and install from URL
 				if ($programInfo['url']) {
 					try {
 						$installerPath = Join-Path $env:TEMP "$selectedProgram.exe"
